+++ src/code_mower/__init__.py
@@ -1,3 +1,3 @@
 """Code Mower package."""
 
-__version__ = "0.5.0a10"
+__version__ = "0.5.0a11"